Move non-handler helper funcs to it's own file
