+++ handler/handler.go
@@ -3,19 +3,15 @@ package handler
 import (
 	"encoding/json"
 	"fmt"
-	"html/template"
 	"io"
 	"io/ioutil"
 	"net/http"
-	"net/url"
 	"reflect"
 	"strings"
 
 	"github.com/Sirupsen/logrus"
 	"github.com/gorilla/mux"
-	"github.com/klarna/eremetic/assets"
 	"github.com/klarna/eremetic/database"
-	"github.com/klarna/eremetic/formatter"
 	"github.com/klarna/eremetic/scheduler"
 	"github.com/klarna/eremetic/types"
 )
@@ -35,20 +31,6 @@ func Create(scheduler types.Scheduler, database database.TaskDB) Handler {
 		scheduler: scheduler,
 		database:  database,
 	}
-}
-
-func absURL(r *http.Request, path string) string {
-	scheme := r.Header.Get("X-Forwarded-Proto")
-	if scheme == "" {
-		scheme = "http"
-	}
-
-	url := url.URL{
-		Scheme: scheme,
-		Host:   r.Host,
-		Path:   path,
-	}
-	return url.String()
 }
 
 // AddTask handles adding a task to the queue
@@ -90,53 +72,12 @@ func (h Handler) AddTask() http.HandlerFunc {
 
 // GetSTDOUT fetches the stdout logs from the agent that ran the task
 func (h Handler) GetSTDOUT() http.HandlerFunc {
-	return h.getFile("stdout")
+	return getFile("stdout")
 }
 
 // GetSTDERR fetches the stderr logs from the agent that ran the task
 func (h Handler) GetSTDERR() http.HandlerFunc {
-	return h.getFile("stderr")
-}
-
-// getFile handles the actual fetching of file from the agent.
-func (h Handler) getFile(file string) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-		taskID := vars["taskId"]
-		task, _ := h.database.ReadTask(taskID)
-
-		if task.SandboxPath == "" {
-			writeJSON(http.StatusNoContent, nil, w)
-			return
-		}
-
-		url := fmt.Sprintf(
-			"http://%s:%d/files/download?path=%s/%s",
-			task.AgentIP,
-			task.AgentPort,
-			task.SandboxPath,
-			file,
-		)
-
-		logrus.WithField("url", url).Debug("Fetching file from sandbox")
-
-		response, err := http.Get(url)
-		if err != nil {
-			logrus.WithError(err).Errorf("Unable to fetch %s from agent %s.", file, task.SlaveId)
-			writeJSON(http.StatusInternalServerError, "Unable to fetch upstream file.", w)
-			return
-		}
-		defer response.Body.Close()
-		body, err := ioutil.ReadAll(response.Body)
-		if err != nil {
-			logrus.WithError(err).Error("Unable to deserialize file.")
-			writeJSON(http.StatusInternalServerError, "Unable to deserialize file.", w)
-			return
-		}
-		w.Header().Set("Content-Type", "text/plain; charset=UTF-8")
-		w.WriteHeader(http.StatusOK)
-		w.Write(body)
-	}
+	return getFile("stderr")
 }
 
 // GetTaskInfo returns information about the given task.
@@ -169,74 +110,4 @@ func (h Handler) ListRunningTasks() http.HandlerFunc {
 		}
 		writeJSON(200, tasks, w)
 	}
-}
-
-func handleError(err error, w http.ResponseWriter, message string) {
-	if err == nil {
-		return
-	}
-
-	errorMessage := ErrorDocument{
-		err.Error(),
-		message,
-	}
-
-	if err = writeJSON(422, errorMessage, w); err != nil {
-		logrus.WithError(err).WithField("message", message).Panic("Unable to respond")
-	}
-}
-
-func writeJSON(status int, data interface{}, w http.ResponseWriter) error {
-	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
-	w.WriteHeader(status)
-	return json.NewEncoder(w).Encode(data)
-}
-
-func renderHTML(w http.ResponseWriter, r *http.Request, task types.EremeticTask, taskID string) {
-	var templateFile string
-
-	data := make(map[string]interface{})
-	funcMap := template.FuncMap{
-		"ToLower":    strings.ToLower,
-		"FormatTime": formatter.FormatTime,
-	}
-
-	if reflect.DeepEqual(task, (types.EremeticTask{})) {
-		templateFile = "error_404.html"
-		data["TaskID"] = taskID
-	} else {
-		templateFile = "task.html"
-		data = makeMap(task)
-	}
-
-	source, _ := assets.Asset(fmt.Sprintf("templates/%s", templateFile))
-	tpl, err := template.New(templateFile).Funcs(funcMap).Parse(string(source))
-
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		logrus.WithError(err).WithField("template", templateFile).Error("Unable to render template")
-		return
-	}
-
-	err = tpl.Execute(w, data)
-}
-
-func makeMap(task types.EremeticTask) map[string]interface{} {
-	data := make(map[string]interface{})
-
-	data["TaskID"] = task.ID
-	data["CommandEnv"] = task.Environment
-	data["CommandUser"] = task.User
-	data["Command"] = task.Command
-	// TODO: Support more than docker?
-	data["ContainerImage"] = task.Image
-	data["FrameworkID"] = task.FrameworkId
-	data["Hostname"] = task.Hostname
-	data["Name"] = task.Name
-	data["SlaveID"] = task.SlaveId
-	data["Status"] = task.Status
-	data["CPU"] = fmt.Sprintf("%.2f", task.TaskCPUs)
-	data["Memory"] = fmt.Sprintf("%.2f", task.TaskMem)
-
-	return data
 }
